TAF-14 Move common things to test-common
