+++ skyxplore-ui-test/build.gradle
@@ -12,7 +12,6 @@ buildscript {
     }
 }
 
-apply plugin: 'java'
 apply plugin: 'io.qameta.allure'
 
 compileJava.options.encoding = "UTF-8"
@@ -123,9 +122,6 @@ dependencies {
     compile group: 'org.seleniumhq.selenium', name: 'selenium-api', version: '3.141.59'
     compile("org.assertj:assertj-core:3.11.1")
     compile 'io.rest-assured:rest-assured:3.3.0'
-    compile group: 'io.qameta.allure', name: 'allure-testng', version: '2.7.0'
-    compile group: 'org.aspectj', name: 'aspectjweaver', version: '1.9.3'
-    compile 'io.qameta.allure:allure-gradle:2.7.0'
     compile group: 'com.github.automatedowl', name: 'allure-environment-writer', version: '1.0.0'
     testCompile group: 'org.awaitility', name: 'awaitility', version: '4.0.0-rc1'
 }
