+++ skyxplore-ui-test/src/test/java/com/mentoring/pageobject/IndexPage.java
@@ -1,6 +1,6 @@
 package com.mentoring.pageobject;
 
-import com.mentoring.model.User;
+import com.mentoring.generator.User;
 import com.mentoring.config.Config;
 import io.qameta.allure.Step;
 import lombok.extern.slf4j.Slf4j;
