+++ skyxplore-ui-test/src/test/java/com/mentoring/test/assertionsteps/CommonAssertions.java
@@ -1,7 +1,7 @@
 package com.mentoring.test.assertionsteps;
 
 import com.mentoring.pageobject.NotificationContainer;
-import com.mentoring.util.Messages;
+import com.mentoring.messages.Messages;
 import org.assertj.core.api.SoftAssertions;
 
 public final class CommonAssertions {
