+++ test-common/build.gradle
@@ -11,7 +11,6 @@ buildscript {
         classpath "io.qameta.allure:allure-gradle:2.7.0"
     }
 }
-apply plugin: 'java'
 
 compileJava.options.encoding = "UTF-8"
 compileTestJava.options.encoding = "UTF-8"
